// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/employee-bonus.ts
export class EmployeeBonus {
    Id: string;
    employeeIds: string[] = [];
    Bonus: number;
    PayrollRunEmployeeId: string;
    IsArchived: boolean;
    employeeId: string;
    IsApproved: boolean;
    EmployeeName: string;
    EmployeeNumber: string;
    type: number;
    amountType: number;
    payRollComponetName: string;
    payRollComponentId: string;
    isCtcType: boolean;
    percentage: number;
    value: number;
    generatedDate: Date;
    profileImage: string;
    userId: string;
    isPaid: boolean;
}